Only auto-own items picked up by controlled minion; Set a dummy controller for dead creatures to avoid crashes; stop ents from disappearing; connect hill areas in level gen; smaller penalty for killing same tribe but different village

+++ village_control.h
@@ -85,7 +85,7 @@ class VillageControl : public CollectiveControl {
   const Location* SERIAL(location);
   vector<Villain> SERIAL(villains);
 
-  map<const Collective*, int> SERIAL(victims);
+  map<const Collective*, double> SERIAL(victims);
 };
 
 #endif
